Backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
